ibd_db(feature): - jtb324 - 11/9/25 - 8:58pm ---
finished script that will load data into a redis database
---
Created a script that can read the ibd segments into a redis database.
It use the pair1 and pair2 ids as a key and then stores the serialized
ibd segment objects as value in a redis-list. This service will be
called at the beginning of the compadre script when it is needed through
a subroutine in perl