feature(main.py): - jtb324 - 11/10/25 - 7:41pm ---
created a fastapi server to server as middleman for redis DB:
---
The legacy perl code used in COMPADRE is a little complicated and can
make connecting to services confusing (and debugging errors and handling
errors even more confusing). If the perl script had to directly query
the redis database then the code would get messy. To alleviate some of
this messinessm I created a fastapi server to act as a middlepoint. It
will accept requests from the perl script, pass them to the redis
database, and then return responses (or errors) to perl as a json
object. This module can handle its own logging.